+++ test/mustache_test.rb
@@ -194,4 +194,60 @@ data
                                                      :server => 'example.com',
                                                      :deploy_to => '/var/www/example.com' )
   end
+
+  def test_reports_type_errors_in_sections
+    instance = Mustache.new
+    instance[:list] = [ :item, 1234 ]
+    instance.template = '{{#list}} <li>{{item}}</li> {{/list}}'
+
+    assert_raise TypeError do
+      instance.render
+    end
+  end
+
+  def test_enumerable_sections_accept_a_hash_as_a_context
+    instance = Mustache.new
+    instance[:list] = { :item => 1234 }
+    instance.template = '{{#list}} <li>{{item}}</li> {{/list}}'
+
+    assert_equal '<li>1234</li>', instance.render.strip
+  end
+
+  def test_knows_when_its_been_compiled_when_set_with_string
+    klass = Class.new(Mustache)
+
+    assert ! klass.compiled?
+    klass.template = 'Hi, {{person}}!'
+    assert klass.compiled?
+  end
+
+  def test_knows_when_its_been_compiled_when_using_a_file_template
+    klass = Class.new(Simple)
+    klass.template_file = File.dirname(__FILE__) + '/../examples/simple.html'
+
+    assert ! klass.compiled?
+    klass.render
+    assert klass.compiled?
+  end
+
+  def test_an_instance_knows_when_its_class_is_compiled
+    instance = Simple.new
+
+    assert ! Simple.compiled?
+    assert ! instance.compiled?
+
+    Simple.render
+
+    assert Simple.compiled?
+    assert instance.compiled?
+  end
+
+  def test_knows_when_its_been_compiled_at_the_instance_level
+    klass = Class.new(Mustache)
+    instance = klass.new
+
+    assert ! instance.compiled?
+    instance.template = 'Hi, {{person}}!'
+    assert instance.compiled?
+  end
 end