+++ HISTORY.md
@@ -4,6 +4,9 @@
   hashes.
 * Added a shortcut: if a section's value is a single hash, treat is as
   a one element array whose value is the hash.
+* Bugfix: String templates set at the class level were not compiled
+* Added a class-level `compiled?` method for checking if a template
+  has been compiled.
 
 ## 0.2.2 (2009-10-11)